pyproject.toml
[tool.poetry]
name = "butcher"
version = "0.1.0"
description = ""
authors = ["Alex Root Junior <jroot.junior@gmail.com>"]
readme = "README.md"
packages = [
    { include = "butcher" }
]

[tool.poetry.scripts]
butcher = 'butcher.shell.main:main'

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
lxml = "^4.9.1"
PyYAML = "^6.0"
autoflake = "^1.7.0"
black = "^22.10.0"
isort = "^5.10.1"
Jinja2 = "^3.1.2"
pydantic = "^1.10.2"
click = "^8.1.3"
rule-engine = "^3.5.0"
devtools = "^0.9.0"
astor = "^0.8.1"
libcst = "^0.4.9"
orjson = "^3.8.1"
alive-progress = "^2.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 99
target-version = ['py310']
exclude = '''
(
    \.eggs
  | \.git
  | \.tox
  | build
  | dist
  | venv
  | docs
)
'''


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 99
